Make it possible to build all packages independently
