+++ packages/chart/package.json
@@ -29,37 +29,53 @@
     "@cdc/core": "^4.25.7",
     "@hello-pangea/dnd": "^16.2.0",
     "@react-spring/web": "^9.7.5",
+    "@rollup/plugin-dsv": "^3.0.2",
+    "@visx/annotation": "^3.3.0",
     "@visx/axis": "3.12.0",
+    "@visx/brush": "^3.12.0",
     "@visx/curve": "3.12.0",
+    "@visx/drag": "^3.12.0",
     "@visx/event": "3.12.0",
     "@visx/glyph": "3.12.0",
     "@visx/gradient": "3.12.0",
+    "@visx/legend": "3.12.0",
     "@visx/marker": "3.12.0",
     "@visx/mock-data": "3.12.0",
+    "@visx/pattern": "^3.0.0",
+    "@visx/responsive": "^2.10.0",
     "@visx/scale": "3.12.0",
     "@visx/shape": "3.12.0",
     "@visx/stats": "3.12.0",
     "@visx/text": "3.12.0",
     "@visx/tooltip": "3.12.0",
+    "@vitejs/plugin-react": "^4.3.4",
+    "chroma-js": "3.1.2",
     "d3-array": "3.2.4",
     "d3-format": "^3.1.0",
     "d3-sankey": "^0.12.3",
     "d3-time-format": "4.1.0",
+    "html-react-parser": "5.2.3",
     "js-base64": "^2.5.2",
+    "papaparse": "5.5.2",
     "react-accessible-accordion": "^3.3.4",
     "react-icons": "5.5.0",
-    "use-debounce": "^6.0.1"
+    "resize-observer-polyfill": "^1.5.1",
+    "use-debounce": "^6.0.1",
+    "vite": "^4.4.11",
+    "vite-plugin-css-injected-by-js": "^2.4.0",
+    "vite-plugin-md": "^0.21.5",
+    "vite-plugin-svgr": "^2.4.0",
+    "whatwg-fetch": "3.6.20"
   },
   "peerDependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
-  "gitHead": "6097de1ff814001880d9ac64bd66becdc092d63c",
+  "gitHead": "9062881d50c824ee6cdd71868bafd016a5e5694d",
   "devDependencies": {
     "@types/d3-array": "^3.2.1",
     "@types/d3-format": "^3.0.4",
     "@types/d3-sankey": "^0.12.4",
-    "@types/d3-time-format": "^4.0.3",
-    "resize-observer-polyfill": "^1.5.1"
+    "@types/d3-time-format": "^4.0.3"
   }
 }
