+++ packages/dashboard/src/helpers/tests/formatConfigBeforeSave.test.ts
@@ -1,35 +1,37 @@
 import { describe, it, expect } from 'vitest'
 import { cleanSharedFilters } from '../formatConfigBeforeSave'
-import { DashboardConfig } from '@cdc/dashboard/src/types/DashboardConfig'
 
 describe('cleanSharedFilters', () => {
   it('should remove unused shared filters', () => {
-    const config: DashboardConfig = {
+    const config: any = {
       dashboard: {
         sharedFilters: [
-          { id: 1, type: 'filter1' },
-          { id: 2, type: 'filter2' },
-          { id: 3, type: 'filter3' },
-          { id: 4, type: 'filter4' }
+          { id: 1, type: 'datafilter' },
+          { id: 2, type: 'datafilter' },
+          { id: 3, type: 'datafilter' },
+          { id: 4, type: 'datafilter' }
         ]
       },
       visualizations: {
         viz1: { type: 'dashboardFilters', sharedFilterIndexes: [0, 3] },
-        viz2: { type: 'chart'}
+        viz2: { type: 'chart' }
       }
     }
 
     cleanSharedFilters(config)
 
-    expect(config.dashboard.sharedFilters).toEqual([{ id: 1, type: 'filter1' }, { id: 4, type: 'filter4' }])
+    expect(config.dashboard.sharedFilters).toEqual([
+      { id: 1, type: 'datafilter' },
+      { id: 4, type: 'datafilter' }
+    ])
   })
 
   it('should retain used shared filters and remove their active state', () => {
-    const config: DashboardConfig = {
+    const config: any = {
       dashboard: {
         sharedFilters: [
-          { id: 1, type: 'filter1', active: true },
-          { id: 2, type: 'filter2', active: true }
+          { id: 1, type: 'datafilter', active: 'test' },
+          { id: 2, type: 'datafilter', active: 'test' }
         ]
       },
       visualizations: {
@@ -41,17 +43,17 @@ describe('cleanSharedFilters', () => {
     cleanSharedFilters(config)
 
     expect(config.dashboard.sharedFilters).toEqual([
-      { id: 1, type: 'filter1' },
-      { id: 2, type: 'filter2' }
+      { id: 1, type: 'datafilter' },
+      { id: 2, type: 'datafilter' }
     ])
   })
 
   it('should remove values from urlfilter type shared filters', () => {
-    const config: DashboardConfig = {
+    const config: any = {
       dashboard: {
         sharedFilters: [
-          { id: 1, type: 'urlfilter', values: [1, 2, 3] },
-          { id: 2, type: 'filter2' }
+          { id: 1, type: 'urlfilter', values: ['1', '2', '3'] },
+          { id: 2, type: 'datafilter' }
         ]
       },
       visualizations: {
@@ -61,8 +63,6 @@ describe('cleanSharedFilters', () => {
 
     cleanSharedFilters(config)
 
-    expect(config.dashboard.sharedFilters).toEqual([
-      { id: 1, type: 'urlfilter' }
-    ])
+    expect(config.dashboard.sharedFilters).toEqual([{ id: 1, type: 'urlfilter' }])
   })
 })
