+++ packages/editor/package.json
@@ -41,5 +41,12 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
+  "devDependencies": {
+    "vite": "^4.4.11",
+    "@vitejs/plugin-react": "^4.3.4",
+    "vite-plugin-css-injected-by-js": "^2.4.0",
+    "vite-plugin-svgr": "^2.4.0",
+    "@rollup/plugin-dsv": "^3.0.2"
+  },
   "gitHead": "6097de1ff814001880d9ac64bd66becdc092d63c"
 }
