+++ packages/editor/src/components/ConfigureTab.tsx
@@ -7,7 +7,7 @@ import CdcWaffleChart from '@cdc/waffle-chart/src/CdcWaffleChart'
 import CdcMarkupInclude from '@cdc/markup-include/src/CdcMarkupInclude'
 
 import ErrorBoundary from '@cdc/core/components/ErrorBoundary'
-import ConfigContext from '../ConfigContext'
+import ConfigContext from '@cdc/core/contexts/EditorContext'
 import MultiDashboardWrapper from '@cdc/dashboard/src/CdcDashboard'
 import CdcDataTable from '../../../data-table/src/CdcDataTable'
