+++ packages/map/package.json
@@ -29,21 +29,35 @@
     "@googlemaps/markerclusterer": "^2.5.3",
     "@hello-pangea/dnd": "^16.2.0",
     "@react-google-maps/api": "^2.20.6",
+    "@rollup/plugin-dsv": "^3.0.2",
+    "@visx/annotation": "^3.3.0",
+    "@visx/drag": "^3.12.0",
     "@visx/geo": "^3.12.0",
+    "@visx/pattern": "^3.0.0",
+    "@vitejs/plugin-react": "^4.3.4",
     "chroma": "0.0.1",
+    "chroma-js": "3.1.2",
     "d3": "^7.9.0",
     "d3-composite-projections": "^1.4.0",
     "d3-selection": "^3.0.0",
     "d3-zoom": "^3.0.0",
+    "html-react-parser": "5.2.3",
     "leaflet": "^1.9.4",
     "lodash.debounce": "^4.0.8",
+    "papaparse": "5.5.2",
     "react-accessible-accordion": "^3.0.1",
+    "resize-observer-polyfill": "^1.5.1",
     "topojson-client": "^3.1.0",
-    "use-debounce": "^10.0.5"
+    "use-debounce": "^10.0.5",
+    "vite": "^4.4.11",
+    "vite-plugin-css-injected-by-js": "^2.4.0",
+    "vite-plugin-md": "^0.21.5",
+    "vite-plugin-svgr": "^2.4.0",
+    "whatwg-fetch": "3.6.20"
   },
   "peerDependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
-  "gitHead": "6097de1ff814001880d9ac64bd66becdc092d63c"
+  "gitHead": "432a2d1acab22915fafe793cb9da1f10318ff793"
 }
