+++ packages/map/src/CdcMapComponent.tsx
@@ -64,7 +64,7 @@ import useResizeObserver from './hooks/useResizeObserver'
 import { VizFilter } from '@cdc/core/types/VizFilter'
 import { getInitialState, mapReducer } from './store/map.reducer'
 import { RuntimeData } from './types/RuntimeData'
-import EditorContext from '@cdc/editor/src/ConfigContext'
+import EditorContext from '@cdc/core/contexts/EditorContext'
 import MapActions from './store/map.actions'
 import _ from 'lodash'
 import useModal from './hooks/useModal'
